Update Cell.java
In this file, my professor K. Walsh provided basic parts of the class as a shell for me to fill in and use for the game. The parts that he provided include cell states: final variable size, isMine, redMine, isRevealed, and coastIsClear. The contents of each cell state including the ones I just listed I created or filled in myself. The rest of this Cell.java class I wrote on my own. My professors comment instructions are left in so it is clear what was asked of me in this project.